Removes `test_functional.py`
Moves a few tests out from `test_functional.py` into more appropriate places
and removes the rest, which were redundant with other existing tests.
